color-gram.py
import colorgram
rgb_colors = []
colors = colorgram.extract('image.jpg', 30)
for color in colors:
    r= color.rgb.r
    b= color.rgb.b
    g= color.rgb.g
    new_color = (r,b,g)
    rgb_colors.append(new_color)

print(rgb_colors)